Rename ValidatorApiClientUtil
