+++ teku/src/main/java/tech/pegasys/teku/cli/subcommand/ValidatorClientCommand.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.teku.cli.subcommand;
 
-import static tech.pegasys.teku.cli.subcommand.ValidatorApiClientUtil.getSpecOrExit;
+import static tech.pegasys.teku.cli.subcommand.RemoteSpecLoader.getSpecOrExit;
 import static tech.pegasys.teku.infrastructure.logging.SubCommandLogger.SUB_COMMAND_LOG;
 
 import java.util.concurrent.Callable;
