+++ teku/src/test/java/tech/pegasys/teku/cli/subcommand/ValidatorClientCommandTest.java
@@ -35,10 +35,10 @@ public class ValidatorClientCommandTest extends AbstractBeaconNodeCommandTest {
             Resources.getResource("tech/pegasys/teku/cli/subcommand/test-spec.yaml").getPath());
     final String[] args = {"vc", "--network", "auto", "--beacon-node-api-endpoint", endpoint};
 
-    try (MockedStatic<ValidatorApiClientUtil> mockValidatorClient =
-        mockStatic(ValidatorApiClientUtil.class)) {
+    try (MockedStatic<RemoteSpecLoader> mockValidatorClient =
+        mockStatic(RemoteSpecLoader.class)) {
       mockValidatorClient
-          .when(() -> ValidatorApiClientUtil.getSpecOrExit(URI.create(endpoint)))
+          .when(() -> RemoteSpecLoader.getSpecOrExit(URI.create(endpoint)))
           .thenReturn(testSpec);
 
       int parseResult = beaconNodeCommand.parse(args);
